+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The project follows [Semantic Versioning 2.0.0](https://semver.org/).
 
-## [Unreleased]
+## [0.3.0] - 2026-04-02
 
 ### Added